+++ modules/memory_manager/builtin/cpp/Gateway.cpp
@@ -28,6 +28,7 @@
 #include "varislockBuiltin.hpp"
 #include "varunlockBuiltin.hpp"
 #include "memoryBuiltin.hpp"
+#include "isvarBuiltin.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
@@ -46,7 +47,7 @@ static const nlsGateway gateway[] =
     { "varunlock", Nelson::MemoryGateway::varunlockBuiltin, 0, 2 },
     { "varislock", Nelson::MemoryGateway::varislockBuiltin, 1, 2 },
     { "memory", Nelson::MemoryGateway::memoryBuiltin, 2, 0 },
-
+	{ "isvar", Nelson::MemoryGateway::isvarBuiltin, 1, -1 },
 };
 //=============================================================================
 NLSGATEWAYFUNC(gateway)